requirements.txt
# Core dependencies
pandas>=2.0.0
pyarrow>=14.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Neo4j
neo4j>=5.14.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
structlog>=23.2.0

# CLI
click>=8.1.0

# Data validation
pandera>=0.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Airflow (for orchestration)
apache-airflow>=2.8.0

# Development
ipython>=8.0.0
jupyter>=1.0.0


